VYBES-92: fix misaligned search icon inside playlist
